Move index.js to its dedicated dir
